+++ packages/sycamore/src/template.rs
@@ -5,7 +5,7 @@ use std::fmt;
 use std::rc::Rc;
 
 use crate::generic_node::GenericNode;
-use crate::rx::{create_memo, StateHandle};
+use crate::reactive::{create_memo, StateHandle};
 
 /// Internal type for [`Template`].
 #[derive(Clone)]
